+++ pkgs/xxhash/src/str/index.ts
@@ -1,3 +1,3 @@
-export { xxh32Str } from "./xxh32Str.js";
-export { xxh64Str } from "./xxh64Str.js";
-export type { xxhStr } from "./xxhStr.js";
+export { xxh32Str } from "./xxh32Str.ts";
+export { xxh64Str } from "./xxh64Str.ts";
+export type { xxhStr } from "./xxhStr.ts";