+++ pkgs/xxhash/vitest.config.ts
@@ -1,7 +1,1 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  test: {
-    isolate: false,
-  },
-});
+../dev/tmpls/vitest.config.tmpl.ts